+++ .env.example
@@ -142,6 +142,8 @@ SENTRY_DSN=
 # Set to true to offload decision rules content to blob storage.
 # This will enable a background task moving the rule payloads that are older than `OFFLOADING_BEFORE`
 # to blob storage to improve database efficiency, while keeping retention requirements.
+# Note that OFFLOADING_BUCKET_URL must be set on the API and worker, while the others need only be set on the worker.
+# OFFLOADING_BUCKET_URL: see section above on object storage configuration.
 # OFFLOADING_ENABLED=false
 # OFFLOADING_JOB_INTERVAL=30m
 # OFFLOADING_BEFORE=168h
