--- models/ast/ast_custom_list_attr.go
@@ -1,23 +1,0 @@
-package ast
-
-// ======= CustomListAccess =======
-
-var AttributeFuncCustomListAccess = struct {
-	FuncAttributes
-	ArgumentCustomListId string
-}{
-	FuncAttributes: FuncAttributes{
-		DebugName: "FUNC_CUSTOM_LIST_ACCESS",
-		AstName:   "CustomListAccess",
-		NamedArguments: []string{
-			"customListId",
-		},
-		Cost: 30,
-	},
-	ArgumentCustomListId: "customListId",
-}
-
-func NewNodeCustomListAccess(customListId string) Node {
-	return Node{Function: FUNC_CUSTOM_LIST_ACCESS}.
-		AddNamedChild(AttributeFuncCustomListAccess.ArgumentCustomListId, NewNodeConstant(customListId))
-}
